requirements.txt
# Python dependencies
openai
PyPDF2
faiss-cpu
numpy